feat(): Split attributes and description of character
