+++ domain/gamedomain/src/main/java/com/angelus/gamedomain/entities/CharacterSkills.kt
@@ -1,0 +1,5 @@
+package com.angelus.gamedomain.entities
+
+data class Skill(val name: String, val level: Int)
+
+data class CharacterSkills(val skills: List<Skill>)
